Extract `IndexViewOptionalFeaturesMixin` from generic `IndexView`
