+++ wagtail/admin/views/generic/__init__.py
@@ -2,8 +2,10 @@ from .base import WagtailAdminTemplateMixin  # noqa
 from .mixins import (  # noqa
     BeforeAfterHookMixin,
     HookResponseMixin,
+    IndexViewOptionalFeaturesMixin,
     LocaleMixin,
     PanelMixin,
+    RevisionsRevertMixin,
 )
 from .models import (  # noqa
     CreateView,
